# Refactor search