--- a/src/Digdir.Domain.Dialogporten.Application/Common/Extensions/DbSetExtensions.cs
+++ b/src/Digdir.Domain.Dialogporten.Application/Common/Extensions/DbSetExtensions.cs
@@ -1,6 +1,7 @@
 using System.Globalization;
 using System.Text;
 using Digdir.Domain.Dialogporten.Application.Externals.AltinnAuthorization;
+using Digdir.Domain.Dialogporten.Application.Features.V1.Common;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,14 +9,24 @@ namespace Digdir.Domain.Dialogporten.Application.Common.Extensions;
 
 public static class DbSetExtensions
 {
-    public static (string sql, object[] parameters) GeneratePrefilterAuthorizedDialogsSql(DialogSearchAuthorizationResult authorizedResources)
+    public static (string sql, object[] parameters) GeneratePrefilterAuthorizedDialogsSql(
+        DialogSearchAuthorizationResult authorizedResources,
+        DeletedFilter deletedFilter)
     {
         var parameters = new List<object>();
+        var deletedFilterCondition = deletedFilter switch
+        {
+            DeletedFilter.Include => "",
+            DeletedFilter.Exclude => "NOT \"Deleted\" AND ",
+            DeletedFilter.Only => "\"Deleted\" AND",
+            _ => throw new ArgumentOutOfRangeException(nameof(deletedFilter), deletedFilter, null)
+        };
+
         var sb = new StringBuilder()
             .AppendLine(CultureInfo.InvariantCulture, $"""
-                SELECT *
+                SELECT "Id", "ServiceResource"
                 FROM "Dialog"
-                WHERE "Id" = ANY(@p{parameters.Count})
+                WHERE {deletedFilterCondition} ("Id" = ANY(@p{parameters.Count})
                 """);
         parameters.Add(authorizedResources.DialogIds);
 
@@ -39,12 +50,17 @@ public static class DbSetExtensions
             parameters.Add(resources);
         }
 
+        sb.AppendLine(")");
+
         return (sb.ToString(), parameters.ToArray());
     }
 
-    public static IQueryable<DialogEntity> PrefilterAuthorizedDialogs(this DbSet<DialogEntity> dialogs, DialogSearchAuthorizationResult authorizedResources)
+    public static IQueryable<DialogEntity> PrefilterAuthorizedDialogs(
+        this DbSet<DialogEntity> dialogs,
+        DialogSearchAuthorizationResult authorizedResources,
+        DeletedFilter? deletedFilter)
     {
-        var (sql, parameters) = GeneratePrefilterAuthorizedDialogsSql(authorizedResources);
+        var (sql, parameters) = GeneratePrefilterAuthorizedDialogsSql(authorizedResources, deletedFilter ?? DeletedFilter.Exclude);
         return dialogs.FromSqlRaw(sql, parameters);
     }
 }