--- a/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/DialogDto.cs
+++ b/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/DialogDto.cs
@@ -59,3 +59,8 @@ public sealed class IntermediateDialogDto : DialogDtoBase
 {
     public List<DialogContent> Content { get; set; } = [];
 }
+
+public sealed class DialogIds
+{
+    public List<Guid> Ids { get; set; } = [];
+}