--- a/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/DialogDtoBase.cs
+++ b/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/DialogDtoBase.cs
@@ -130,7 +130,7 @@ public class DialogDtoBase
     /// <summary>
     /// The latest entry in the dialog's activity log.
     /// </summary>
-    public DialogActivityDto? LatestActivity { get; set; }
+    public DialogActivityDto LatestActivity { get; set; } = null!;
 
     /// <summary>
     /// The list of seen log entries for the dialog newer than the dialog ChangedAt date.