--- a/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/SearchDialogQuery.cs
+++ b/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Queries/Search/SearchDialogQuery.cs
@@ -174,7 +174,13 @@ internal sealed class SearchDialogQueryHandler : IRequestHandler<SearchDialogQue
         var resourceIds = await _userResourceRegistry.GetCurrentUserResourceIds(cancellationToken);
         var searchExpression = Expressions.LocalizedSearchExpression(request.Search, request.SearchLanguageCode);
 
-        var dialogQuery = _db.Dialogs.AsQueryable();
+        var formattedServiceOwnerLabels = request.ServiceOwnerLabels?
+            .Select(label => label.EndsWith("*", StringComparison.OrdinalIgnoreCase)
+                ? label.TrimEnd('*').ToLower(CultureInfo.InvariantCulture) + "%"
+                : label.ToLower(CultureInfo.InvariantCulture))
+            .ToList();
+
+        var dialogIdsQuery = _db.Dialogs.AsQueryable();
 
         // If the service owner impersonates an end user, we need to filter the dialogs
         // based on the end user's authorization, not the service owner's (which is
@@ -185,20 +191,11 @@ internal sealed class SearchDialogQueryHandler : IRequestHandler<SearchDialogQue
                 request.Party ?? [],
                 request.ServiceResource ?? [],
                 cancellationToken);
-            dialogQuery = _db.Dialogs.PrefilterAuthorizedDialogs(authorizedResources);
+            dialogIdsQuery = _db.Dialogs.PrefilterAuthorizedDialogs(authorizedResources, request.Deleted);
         }
 
-        var formattedServiceOwnerLabels = request.ServiceOwnerLabels?
-            .Select(label => label.EndsWith("*", StringComparison.OrdinalIgnoreCase)
-                ? label.TrimEnd('*').ToLower(CultureInfo.InvariantCulture) + "%"
-                : label.ToLower(CultureInfo.InvariantCulture))
-            .ToList();
-
-        var paginatedList = await dialogQuery
-            .Include(x => x.Content)
-                .ThenInclude(x => x.Value.Localizations)
-            .Include(x => x.ServiceOwnerContext)
-                .ThenInclude(x => x.ServiceOwnerLabels)
+        // Apply all original filtering criteria to the ID query.
+        dialogIdsQuery = dialogIdsQuery
             .WhereIf(!request.ServiceResource.IsNullOrEmpty(),
                 x => request.ServiceResource!.Contains(x.ServiceResource))
             .WhereIf(!request.Party.IsNullOrEmpty(), x => request.Party!.Contains(x.Party))
@@ -221,15 +218,33 @@ internal sealed class SearchDialogQueryHandler : IRequestHandler<SearchDialogQue
                 x.Content.Any(x => x.Value.Localizations.AsQueryable().Any(searchExpression)) ||
                 x.SearchTags.Any(x => EF.Functions.ILike(x.Value, request.Search!))
             )
-            .WhereIf(request.Deleted == DeletedFilter.Exclude, x => !x.Deleted)
-            .WhereIf(request.Deleted == DeletedFilter.Only, x => x.Deleted)
+            // If we have enduserid, we have already filtered out the deleted dialogs
+            .WhereIf(request.EndUserId is null && request.Deleted == DeletedFilter.Exclude, x => !x.Deleted)
+            .WhereIf(request.EndUserId is null && request.Deleted == DeletedFilter.Only, x => x.Deleted)
             .WhereIf(formattedServiceOwnerLabels is not null && formattedServiceOwnerLabels.Count != 0, x =>
                 formattedServiceOwnerLabels!
                     .All(formattedLabel =>
                         x.ServiceOwnerContext.ServiceOwnerLabels
                             .Any(l => EF.Functions.ILike(l.Value, formattedLabel))))
             .WhereIf(request.ExcludeApiOnly == true, x => !x.IsApiOnly)
-            .Where(x => resourceIds.Contains(x.ServiceResource))
+            // TODO! If we have a supplied ServiceResource filter, we should only check if the resourcesIds contains the supplied ServiceResources
+            .Where(x => resourceIds.Contains(x.ServiceResource)) // Final mandatory filter
+            .IgnoreQueryFilters();
+
+        // Now, apply ordering and pagination to the ID query and execute it.
+        var dialogIds = await dialogIdsQuery
+            .Select(x => x.Id) // Select only the primary key
+            .Take(request.Limit ?? 100)
+            .ToListAsync(cancellationToken);
+        //.ProjectTo<IntermediateDialogDto>(_mapper.ConfigurationProvider)
+        //.ToPaginatedListAsync(request, cancellationToken: cancellationToken);
+
+        var paginatedList = await _db.Dialogs
+            .Where(x => dialogIds.Contains(x.Id))
+            .Include(x => x.Content)
+                .ThenInclude(x => x.Value.Localizations)
+            .Include(x => x.ServiceOwnerContext)
+                .ThenInclude(x => x.ServiceOwnerLabels)
             .IgnoreQueryFilters()
             .ProjectTo<IntermediateDialogDto>(_mapper.ConfigurationProvider)
             .ToPaginatedListAsync(request, cancellationToken: cancellationToken);