--- a/src/Digdir.Domain.Dialogporten.Infrastructure/InfrastructureExtensions.cs
+++ b/src/Digdir.Domain.Dialogporten.Infrastructure/InfrastructureExtensions.cs
@@ -62,6 +62,7 @@ public static class InfrastructureExtensions
                 {
                     o.UseQuerySplittingBehavior(QuerySplittingBehavior.SplitQuery);
                 })
+                .EnableSensitiveDataLogging()
                 .AddInterceptors(
                     services.GetRequiredService<PopulateActorNameInterceptor>(),
                     services.GetRequiredService<ConvertDomainEventsToOutboxMessagesInterceptor>()