This attempts to adress several issues with todays search
- As these queries must be split anyway to avoid cartesian explosions, the handler is changed to initially fetch a minimal list of dialog ids (without any joins, but with all filters). The resulting dialog-id list is then used as input for subsequent calls.
- The automapper profile for DialogEntity->IntermediateDialogDto is simplified, and mapping logic is moved to IntermediateDialogDto->DialogDto. This creates much simpler SQL which appears to perform better. 
- The deleted-query filter is ignored, and deleted handling is moved to dialog Id-materialization step

Local testing (with appx 1M dialogs) seems to reduce cold service owner requests filtering on party from ~600ms to 350ms, hot lookups from ~100ms to ~60ms. This _appears_ to be due to CPU-wins in postgresql. I/O overhead for cold requests are still assumed to dominate time spent. 

Issues
- Sorting and pagination is broken (not sure how to fix it, pagination code is very opaque)
- ServiceOwnerContext and EnduserContext is JOINed for every sub-query (not clear why)
- Only implemented in serviceowner (easy to port to enduser when done)

Further improvements
- Intersect serviceowner resourceIds when supplying enduser id to improve performance of dialog id materialization
- Partial index with Deleted
- Create index with INCLUDING on the timestamp columns, deleted, serviceowner and party. This will cause the initial id-loopkup to be a index-only scan, avoiding the heap for even cold lookups
- Change default sort order i API to match arbeidsflate (contentupdatedat DESC)
- Cluster all immediate children of Dialog on DialogId